refactor: cache 를 위한 해시기능 클래스를 cache 에서 util.hash 로 이동
